refactor(headless-future): internalize converse request builder types
