+++ packages/headless-future/src/api/index.ts
@@ -10,20 +10,3 @@
 
 // Search API
 export {executeSearchAPI} from './search/search.js';
-
-// HTTP utilities (for future API clients)
-export {executeHttpRequest} from './shared/http.js';
-export type {HttpRequestOptions, HttpResponse} from './shared/http.js';
-
-// Stream utilities (for future conversational clients)
-export {readEventStream} from './shared/stream.js';
-export type {ReadEventStreamOptions} from './shared/stream.js';
-
-// Conversation request utilities
-export {
-  buildConverseRequestBody,
-  type ConverseRequestBody,
-} from './conversation/build-converse-request-body.js';
-
-// Error handling utilities (for future API clients)
-export {transformError, isSuccessResponse} from './shared/error-handling.js';
